refactor: split App into components refactor: update readme
